[WS] Add diabetes and heart attack prediction endpoints

+++ backend/src/main/java/com/healthapp/backend/model/User.java
@@ -1,6 +1,12 @@
 package com.healthapp.backend.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -28,7 +34,7 @@ public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "username", unique = true, nullable = false)
     private String username;
